Parameterised Controller on the type of register with which it can work.
While adding the PWM controller I realised that registers for that controller will be specified as String, not as Byte. So this is an attempt to parameterise the type of registers. Not entirely happy that controllers now are bound to a register type, but it also makes some sense as the bus medium (eg. for I2C and SPI) does dictate what format the register value can take on.