Refactors the SPI metrics out of the SPI-to-TCP utility and into the SPI controller
NB. that this is the first time state has appeared in the SPI controller (or
any of the controllers). This was a difficult decision to make. The
properly FP way of doing this (with fs2) would be to pipe the SPI events
to a stream that then aggregates them and derives the downstream
results. This has real downsides:
* it binds the SPI controller to fs2.
* alternatively, it could live as an fs2-extension to the SPI controller
but; then the metrics are not a piece of core functionality, and I think
they should be.

It is my intent that consumers of this information will access it in an
FP-ish way, eg. using fs2 to manage the threading & polling frequency.