Adds an fs2 wrapper around the GPS comms and message processing
Originally it was intended to have a pure/non-fs2 equivalent of the
`Device` for the GPS and the provide fs2 convenience wrappers for this.
However it was discovered that providing a sensible API to a GPS would
require either maintaining state in-object or having an API that would
still require the caller to maintain state between calls (no matter
what the API there is no guarantee that a data exchange won't end
mid-message).

And so instead there is an fs2 equivalent of `Device` for the GPS which
removes the state management from the caller.