Extracts management of metrics _back_ out of the SpiController
I keep going back and forth on this. While the details of _events_ that
are of interest are entirely coupled to the generator of said events,
the ways in which those events become interesting metrics probably
varies a lot based on the consumer.

As a result for the moment the interpretation of SPI events is coming
out of the SpiController and into SpiToTcp utility.